5/19 初次完成第一版

+++ src/main/java/com/pitaya/smart_rest/system/entity/Org.java
@@ -31,7 +31,7 @@ public class Org implements Serializable {
     private Integer id;
     private String orgName;
 
-    private Long parentId;
+    private Integer parentId;
 
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "GMT+8")
     private Date createDate;
